+++ smallworld/__init__.py
@@ -8,7 +8,7 @@ __author__ = metadata["Author"]
 __version__ = metadata["version"]
 
 from . import analyses, cpus, emulators, exceptions, hinting, initializers, state
-from .helpers import analyze, emulate
+from .helpers import analyze, emulate, fuzz
 from .utils import setup_hinting, setup_logging
 
 __all__ = [
@@ -19,8 +19,9 @@ __all__ = [
     "hinting",
     "initializers",
     "state",
-    "analyze",
-    "emulate",
     "setup_hinting",
     "setup_logging",
+    "analyze",
+    "emulate",
+    "fuzz",
 ]