+++ smallworld/analyses/__init__.py
@@ -1,4 +1,12 @@
 from .analysis import Analysis, Filter
+from .code_coverage import CodeCoverage
+from .control_flow_tracer import ControlFlowTracer
 from .input_colorizer import InputColorizerAnalysis
 
-__all__ = ["Analysis", "Filter", "InputColorizerAnalysis"]
+__all__ = [
+    "Analysis",
+    "Filter",
+    "InputColorizerAnalysis",
+    "ControlFlowTracer",
+    "CodeCoverage",
+]