+++ smallworld/hinting.py
@@ -27,11 +27,34 @@ class Hint(utils.Serializable):
 
 
 @dataclass(frozen=True)
+class ControlFlowHint(Hint):
+    """Represents control flow going from the from_instruction to the to_instruction.
+
+    Arguments:
+        from_instruction: The from instruction
+        to_instruction: The to instruction
+    """
+
+    from_instruction: typing.Any
+    to_instruction: typing.Any
+
+
+@dataclass(frozen=True)
+class CoverageHint(Hint):
+    """Hold the a map of program counter to hit counter for an execution
+
+    Arguments:
+        coverage: A map from program counter to hit count
+    """
+
+    coverage: typing.Dict[int, int]
+
+
+@dataclass(frozen=True)
 class EmulationException(Hint):
     """Something went wrong emulating this instruction"""
 
     instruction: typing.Any
-    micro_exec_num: int
     instruction_num: int
     exception: str